JVMCBC-1001: Mark Endpoint as free() on (re)connect.
Motivation
----------
There is a possibility that if a channel goes into an inactive
state before the full response has been decoded that the underlying
channel will be re-connected and ready to go again, but the outer
endpoint is not in a free() state - which only happens once the
request is fully decoded right now.

Modifications
-------------
At the moment it is unclear why a channel would become inactive/
would be reconnected before the full response has been decoded,
but we need to make sure to handle that possibility.

Per definition, when a endpoint connects fresh there is no
request in the pipeline so it is free() (which is also the initial
state).

So this change marks an endpoint as free during connect and reconnect.
Most of the time it will be free anyways, but should it not this change
will make sure.

Change-Id: I53d8bd1c619e9fd64a4943f1ce94a3fe5c57ae0a
Reviewed-on: http://review.couchbase.org/c/couchbase-jvm-core/+/158342
Reviewed-by: David Nault <david.nault@couchbase.com>
Tested-by: Michael Nitschinger <michael.nitschinger@couchbase.com>